backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Pillow==10.2.0
python-dotenv==1.0.1
Werkzeug==2.3.7
firebase-admin==6.2.0
gunicorn
numpy==1.26.4
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-magic==0.4.27
aiofiles==23.2.1
SQLAlchemy==2.0.28
pydantic==2.6.1
pypdf==4.1.0
opencv-python==4.9.0.80
scikit-image==0.22.0
audioread==3.0.0
moviepy==1.0.3
diff-match-patch==20231111
tika==1.24
lxml==4.9.4
requests==2.31.0
matplotlib==3.8.3
seaborn==0.13.2
brotli==1.1.0
langchain==0.1.13
openai==1.16.2
python-docx==1.1.0
# For sensitive mode (NLP)
spacy==2.3.7
textblob==0.17.1
nltk==3.8.1